requirements.txt
dropbox==11.36.2
python-dotenv==1.0.0
requests==2.31.0
slack-sdk==3.21.3
schedule==1.2.0
Flask==2.3.3
gunicorn==21.2.0